backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
anthropic==0.7.1
python-docx==0.8.11
pypdf==3.17.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
nltk==3.8.1
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
cors==1.0.1
